+++ pyriemann/classification.py
@@ -8,6 +8,7 @@ from sklearn.utils.extmath import softmax
 from sklearn.linear_model import LogisticRegression
 from sklearn.pipeline import make_pipeline
 from joblib import Parallel, delayed
+import warnings
 
 from .utils.kernel import kernel
 from .utils.mean import mean_covariance, mean_power
@@ -519,7 +520,7 @@ class KNearestNeighbor(MDM):
 
         return self
 
-    def predict(self, covtest):
+    def predict(self, X=None, covtest=None):
         """Get the predictions.
 
         Parameters
@@ -532,7 +533,13 @@ class KNearestNeighbor(MDM):
         pred : ndarray of int, shape (n_matrices,)
             Predictions for each matrix according to the closest centroid.
         """
-        dist = self._predict_distances(covtest)
+        if covtest is not None:
+            warnings.warn("Input covtest has been renamed into X and will be "
+                          "removed in 0.8.0.", category=DeprecationWarning)
+            assert (X is None)
+            X = covtest
+
+        dist = self._predict_distances(X)
         neighbors_classes = self.classmeans_[np.argsort(dist)]
         pred = _mode_2d(neighbors_classes[:, 0:self.n_neighbors], axis=1)
         return pred